Address Matko's comments
