+++ hazelcast-client/src/test/java/com/hazelcast/client/flakeidgen/FlakeIdGenerator_ClientIntegrationTest.java
@@ -21,8 +21,8 @@ import com.hazelcast.client.config.FlakeIdGeneratorConfig;
 import com.hazelcast.client.test.TestHazelcastFactory;
 import com.hazelcast.concurrent.flakeidgen.FlakeIdConcurrencyTestUtil;
 import com.hazelcast.core.FlakeIdGenerator;
-import com.hazelcast.core.FlakeIdGenerator.IdBatch;
 import com.hazelcast.core.HazelcastInstance;
+import com.hazelcast.core.IdBatch;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.ParallelTest;
 import com.hazelcast.test.annotation.QuickTest;
