+++ hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/flakeidgenerator/NewIdBatchMessageTask.java
@@ -21,7 +21,7 @@ import com.hazelcast.client.impl.protocol.codec.FlakeIdGeneratorNewIdBatchCodec;
 import com.hazelcast.client.impl.protocol.task.AbstractMessageTask;
 import com.hazelcast.concurrent.flakeidgen.FlakeIdGeneratorProxy;
 import com.hazelcast.concurrent.flakeidgen.FlakeIdGeneratorService;
-import com.hazelcast.core.FlakeIdGenerator.IdBatch;
+import com.hazelcast.core.IdBatch;
 import com.hazelcast.instance.Node;
 import com.hazelcast.nio.Connection;
 import com.hazelcast.security.permission.ActionConstants;
